auto-yasnippet.el (aya-create): Add optional args
This commit adds optional arguments BEG and END to `aya-create'. These
arguments allow the to explicitly set the region for `aya-create' to
work on.

Since the arguments are optional, and the function took no arguments
previously, this change is fully backwards-compatible.

The new arguments give more flexibility to customize auto-yasnippet,
e.g. when integrating with evil-mode.

Fixes #30